fix(blorktools): landing page file drop
